+++ src/Data/Export/Exception/NoValueToExportException.php
@@ -1,0 +1,16 @@
+<?php
+declare(strict_types=1);
+
+namespace Remorhaz\JSON\Data\Export\Exception;
+
+use LogicException;
+use Throwable;
+
+final class NoValueToExportException extends LogicException implements ExceptionInterface
+{
+
+    public function __construct(Throwable $previous = null)
+    {
+        parent::__construct("No value to export", 0, $previous);
+    }
+}