+++ src/Processor/Mutator/Exception/ExceptionInterface.php
@@ -1,0 +1,10 @@
+<?php
+declare(strict_types=1);
+
+namespace Remorhaz\JSON\Path\Processor\Mutator\Exception;
+
+use Remorhaz\JSON\Path\Processor\Exception\ExceptionInterface as ProcessorExceptionInterface;
+
+interface ExceptionInterface extends ProcessorExceptionInterface
+{
+}