+++ src/Processor/Mutator/MutatorInterface.php
@@ -10,4 +10,10 @@ interface MutatorInterface
 {
 
     public function deletePaths(NodeValueInterface $rootNode, PathInterface ...$paths): ?NodeValueInterface;
+
+    public function replacePaths(
+        NodeValueInterface $rootNode,
+        NodeValueInterface $newNode,
+        PathInterface ...$paths
+    ): NodeValueInterface;
 }